Use secondary IMU for OSD heading and AHI

+++ src/main/drivers/accgyro/accgyro_bno055.h
@@ -34,5 +34,5 @@ typedef struct {
 
 bool bno055Init(void);
 fpVector3_t bno055GetEurlerAngles(void);
-void bno055FetchEulerAngles(int32_t * buffer);
+void bno055FetchEulerAngles(int16_t * buffer);
 bno055CalibStat_t bno055GetCalibStat(void);
